+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+ - `MetricWrapper.label_names` are now `Set` instead of `List`
  - `StoreOperationDecorator` no longer raises `ValueError` when input of `operation` is not valid, 
 now it is logged as warning and record is ignored
  - `Profiler.delete_metric` and `Profiler.delete_handler`no longer raising `KeyError`when metric/handler not exists